lab3-code/requirements.txt
# Core dependencies for HR Agent with LangGraph
langgraph>=0.0.20
langchain>=0.1.0
langchain-openai>=0.0.5

# HTTP client for API calls
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0

# Optional: For graph visualization
# graphviz>=0.20.1
# ipython>=8.12.0

# Optional: For enhanced async support
# aiohttp>=3.9.0